hybrid-inference-orchestrator/requirements.txt
mcp
google-cloud-run
google-cloud-monitoring
google-cloud-storage
pydantic>=2.0
